Add HSL blend modes
The new modes are HSL Hue, HSL Saturation, HSL Luminosity and HSL Color.
I particularly want HSL Color so I can paint just the lightness first
and then worry about the actual colors later. Most other drawing
programs have these layer modes and the SVG spec supports them as well,
so adding these seems like a sensible feature overall.

I don't know why exactly HSL is used instead of HSV, but I decided to
just stick with what other programs are doing. It would be possible to
implement HSV variants of these blend modes as well (HSV and HSL hue are
identical, but saturation, lightness/value and color are slightly
different), but I don't really see the point of cluttering the layer
modes drop-down with so many modes of questionable usefulness.

HSL Hue applies the hue of the new pixel to the old pixel.

HSL Saturation applies the saturation of the new pixel to the old pixel.

HSL Luminosity applies the lightness value from the old pixel to the new
pixel.

HSL Color applies both the hue and the saturation of the old pixel to
the new pixel. This allows for doing shades and colors separately.