Implement Distort transformation on selections
Also known as "Free Deform" in Paint Tool SAI or "Free Transform" in
Clip Studio. Moving the corners of the selection around and making a
perspective transform, basically.  Moving one of the edges around
basically just moves the adjacent corners simultaneously.

Moving the corners into weird positions currently still causes really
weird transformations to happen that smear the contents of the selection
across the entire canvas. Maybe that could be constrained so it doesn't
happen, or maybe it's just a case of not poking where it hurts.

Relates to #609.