Implement basics for dynamic undo depth
An undo depth change message and the handling thereof. Makes the undo
depth a dynamic value that gets changed by the message. When the undo
depth is changed, no undos are possible beyond its point, providing a
hard cutoff where it switches to the new value.

This implementation is still missing pretty much all client and server
infrastructure, which will be added in the following commits.